Updates for reordering tabs
When user selects a radio button, new arrow buttons appear at the top of the popup window. When pressed, the selected tab moves either up or down. If tab is in the first position, clicking the up arrow does nothing. If it's in the last position, the down arrow does nothing.

If the selected tab is deleted, there is no longer a selected tab, so the arrow buttons disappear.