Reorder tabs (#33)
* New "Reorder" column and supporting elements

New column of radio buttons, new action buttons at the top of the page for moving the selected tab, and various classes in places to support the reordering solution.

Also set the width of some columns to give more space where needed or ensure the smallest size necessary where space is not needed.

* Updates to support reordering tabs

Also widened popup from 40rem to 50rem

* Updates for reordering tabs

When user selects a radio button, new arrow buttons appear at the top of the popup window. When pressed, the selected tab moves either up or down. If tab is in the first position, clicking the up arrow does nothing. If it's in the last position, the down arrow does nothing.

If the selected tab is deleted, there is no longer a selected tab, so the arrow buttons disappear.

* Update popup.css to support single delete action button

Also shrunk the overall width to 45rem from 50rem (was originally 40rem before this branch)

* Update popup.html to support single delete action button

* Update popup.js to support single delete action button

---------

Co-authored-by: Alon Waisman <mowalon@yahoo.com>